pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ollama-forge"
version = "0.1.0"
description = "Batteries-included local AI framework for Ollama — hardware optimization, context compression, multi-agent orchestration, and MCP integration"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
keywords = ["ollama", "local-ai", "agents", "mcp", "rocm", "amd"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "requests>=2.31.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "duckduckgo-search>=6.0.0",
]

[project.optional-dependencies]
tui = ["textual>=0.40.0"]
web = ["fastapi>=0.100.0", "uvicorn>=0.23.0", "jinja2>=3.1.0"]
all = ["ollama-forge[tui,web]"]
dev = ["pytest>=7.0.0", "ollama-forge[all]"]

[project.scripts]
forge = "forge.cli:main"

[project.urls]
Homepage = "https://github.com/ollama-forge/ollama-forge"
Documentation = "https://github.com/ollama-forge/ollama-forge/tree/main/docs"
Repository = "https://github.com/ollama-forge/ollama-forge"
Issues = "https://github.com/ollama-forge/ollama-forge/issues"

[tool.setuptools.packages.find]
include = ["forge*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
